move command to driver
